feat: refactoring carbon backfill to integrate with backfill all
