+++ src/cost_export/carbonExport.py
@@ -1,10 +1,19 @@
 import json
+from typing import Dict
 import requests
 import logging
 from datetime import datetime, timezone, timedelta
 from common import Config
+from api.tokens import TokenManager
 import pyarrow.fs as fs
 from api.s3Api import getS3FileSystem
+from api.carbonS3Api import (
+    carbon_export_backfill_lock_exists,
+    carbon_export_backfill_lock_create,
+)
+from billing import (
+    extract_subscription_ids_from_billing_scope,
+)
 
 logger = logging.getLogger("cost_export")
 
@@ -354,3 +363,99 @@ def save_carbon_data_to_s3(data, file_name, force_overwrite=False):
         logger.error(f"Error saving carbon data to S3: {str(e)}")
         raise
 
+def carbon_emissions_backfill_imp(start_date: datetime, skip_existing: bool = True, force_overwrite: bool = False, write_empty_object: bool = True) -> Dict[int, int]:
+    try:
+        # first check if carbon export lock object exists
+        if carbon_export_backfill_lock_exists():
+            logger.info(f"Carbon Export backfill lock object exists. Skipping Carbon Export backfill")
+            return (0, 0)
+
+        logger.info(f"carbon_emissions_backfill_imp: start_date({start_date}), skip_existing({skip_existing}), force_overwrite({force_overwrite}), write_empty_object({write_empty_object})")
+        current_year, current_month = start_date.year, start_date.month
+
+        # Get access token using managed identity
+        token = TokenManager().azure_token
+        logger.debug(f"cost_mgmt_export_exists: token: {token}")
+        
+        headers = {
+            "Authorization": f"Bearer {token}",
+            "Content-Type": "application/json"
+        }
+        
+        # Extract subscription IDs from billing scope
+        subscription_ids = extract_subscription_ids_from_billing_scope(Config.billing_scope)
+        
+        logger.info(f"Starting carbon backfill for {len(subscription_ids)} subscriptions from {start_date.strftime('%Y-%m-%d')}")
+        
+        processed_months = 0
+        skipped_months = 0
+
+        # backfill should through until the latest carbon export data is available
+        #  which from "carbon_emissions_exporter" is the around the 19th for the last month
+        carbon_api_fetch_date = carbon_export_api_latest_fetch_date()
+        latest_fetch_month, latest_fetch_year = carbon_api_fetch_date.month, carbon_api_fetch_date.year
+        
+        # Process months before API range (create empty records)
+        while (current_year, current_month) <= (latest_fetch_year, latest_fetch_month):
+            month_date = datetime(current_year, current_month, 1, tzinfo=timezone.utc)
+            month_str = month_date.strftime("%Y-%m-01")
+            file_name = f"carbon-emissions-{month_date.strftime('%Y-%m')}.json"
+            
+            # Check if data already exists
+            if skip_existing:
+                exists, existing_path = check_carbon_data_exists(file_name)
+                if exists:
+                    logger.info(f"Skipping {month_str} - data already exists at {existing_path}")
+                    skipped_months += 1
+                    # Move to next month
+                    if current_month == 12:
+                        current_year += 1
+                        current_month = 1
+                    else:
+                        current_month += 1
+                    continue
+            
+            logger.info(f"Processing month: {month_str}")
+
+            # attempt to fetch the Carbon Data from API. if there is no carbon data for that month, the API
+            #  will return an empty "value" array
+            success, emissions_data, error_message = make_carbon_api_request_batched(
+                headers, subscription_ids, month_str
+            )
+
+            if success:
+                if success and len(emissions_data["value"]) == 0:
+                    # Create empty carbon data for months outside API range
+                    emissions_data = empty_emissions_data(month_str)
+                
+                save_carbon_data_to_s3(emissions_data, file_name, force_overwrite=force_overwrite)
+                processed_months += 1
+            else:
+                logger.error(error_message)
+                if write_empty_object:
+                    logger.warning(f"Unable to process month: {month_str}. Writing empty results")
+                    emissions_data = empty_emissions_data(month_str)
+                    save_carbon_data_to_s3(emissions_data, file_name, force_overwrite=force_overwrite)
+                    processed_months += 1
+                else:
+                    logger.error(f"Unable to process month: {month_str}. Not writing empty results")
+                    skipped_months += 1
+            
+            # Move to next month
+            if current_month == 12:
+                current_year += 1
+                current_month = 1
+            else:
+                current_month += 1
+        
+        # gets this far if having processed all carbon export backfill
+        carbon_export_backfill_lock_create()
+
+        return (
+            processed_months,
+            skipped_months,
+        )
+        
+    except Exception as e:
+        error_msg = f"Error in carbon emissions backfill impl: {str(e)}"
+        raise Exception(error_msg)
